# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Virtual Bank API",
      "version": "2.1"
    },
    "paths": {
...
      "/products/{product_id}": {
        "get": {
          "tags": [
            "5 Каталог продуктов"
          ],
          "summary": "2. Получить детали продукта",
          "description": "## 📋 Детальная информация о продукте\n\nПолучение подробных условий конкретного банковского продукта.\n\n### Пример запроса:\n```bash\nGET /products/prod-vb-deposit-001\n```\n\n### Когда использовать:\n- После выбора продукта из каталога\n- Перед созданием лида или заявки\n- Для отображения деталей на странице продукта\n\n### Полезно для UI:\n- Показать полные условия\n- Вычислить предварительную сумму выплат по кредиту\n- Сравнить с другими продуктами",
          "operationId": "get_product_products__product_id__get",
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Product Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/product-agreements/{agreement_id}": {
        "get": {
          "tags": [
            "7 Договоры с продуктами"
          ],
          "summary": "3. Получить детали договора",
          "description": "Получить детали договора\n\n### 🔑 Аутентификация:\n- **bank_token**: укажите `client_id` в query параметре\n- **client_token**: `client_id` определится автоматически\n\n### Для межбанковых запросов:\n- Требуется заголовок `x-product-agreement-consent-id`\n- Требуется заголовок `x-requesting-bank`\n- Согласие должно иметь разрешение `read_product_agreements`",
          "operationId": "get_agreement_product_agreements__agreement_id__get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "agreement_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Agreement Id"
              }
            },
            {
              "name": "client_id",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID клиента (обязательно для bank_token)",
                "title": "Client Id"
              },
              "description": "ID клиента (обязательно для bank_token)",
              "example": "team200-1"
            },
            {
              "name": "x-product-agreement-consent-id",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "X-Product-Agreement-Consent-Id"
              }
            },
            {
              "name": "x-requesting-bank",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "X-Requesting-Bank"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "7 Договоры с продуктами"
          ],
          "summary": "4. Закрыть договор",
          "description": "Закрыть договор\n\n- Депозит: закрыть счет\n- Кредит: погасить задолженность из указанного счета\n- Карта: заблокировать\n\n### 🔑 Аутентификация:\n- **bank_token**: укажите `client_id` в query параметре\n- **client_token**: `client_id` определится автоматически\n\n### Для межбанковых запросов:\n- Требуется заголовок `x-product-agreement-consent-id`\n- Требуется заголовок `x-requesting-bank`\n- Согласие должно иметь разрешение `close_product_agreements`",
          "operationId": "close_agreement_product_agreements__agreement_id__delete",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "agreement_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Agreement Id"
              }
            },
            {
              "name": "client_id",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID клиента (обязательно для bank_token)",
                "title": "Client Id"
              },
              "description": "ID клиента (обязательно для bank_token)",
              "example": "team200-1"
            },
            {
              "name": "x-product-agreement-consent-id",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "X-Product-Agreement-Consent-Id"
              }
            },
            {
              "name": "x-requesting-bank",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "X-Requesting-Bank"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CloseAgreementRequest"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Request"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/product-agreement-consents/request": {
        "post": {
          "tags": [
            "6 Согласия на управление договорами"
          ],
          "summary": "1. Создать запрос согласия на управление договорами",
          "description": "## 📋 Создание запроса на согласие для управления договорами\n\n**OpenBanking Russia Products API Extension**\n\n### 🔑 Аутентификация:\n- **bank_token** (type=\"team\"): укажите `client_id` в query параметре\n- **client_token** (type=\"client\"): `client_id` определится автоматически\n\n### Use Case:\nФинансовый агрегатор или маркетплейс хочет:\n- Показывать все продукты клиента из разных банков\n- Открывать депозиты/карты от имени клиента\n- Закрывать неактуальные договоры\n\n### Пример запроса:\n```json\n{\n  \"requesting_bank\": \"team200\",\n  \"client_id\": \"team200-1\",\n  \"read_product_agreements\": true,\n  \"open_product_agreements\": true,\n  \"close_product_agreements\": false,\n  \"allowed_product_types\": [\"deposit\", \"card\"],\n  \"max_amount\": 1000000.00,\n  \"valid_until\": \"2025-12-31T23:59:59\",\n  \"reason\": \"Финансовый агрегатор для управления продуктами\"\n}\n```\n\n### Разрешения:\n- `read_product_agreements`: просмотр списка договоров (депозиты, кредиты, карты)\n- `open_product_agreements`: открытие новых продуктов\n- `close_product_agreements`: закрытие существующих договоров\n\n### Ограничения:\n- `allowed_product_types`: только указанные типы продуктов\n- `max_amount`: макс сумма для открытия одного продукта",
          "operationId": "create_product_agreement_consent_request_product_agreement_consents_request_post",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "client_id",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID клиента (обязательно для bank_token)",
                "title": "Client Id"
              },
              "description": "ID клиента (обязательно для bank_token)",
              "example": "team200-1"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductAgreementConsentRequestData"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "title": "Response Create Product Agreement Consent Request Product Agreement Consents Request Post"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/product-agreement-consents/{consent_id}": {
        "get": {
          "tags": [
            "6 Согласия на управление договорами"
          ],
          "summary": "2. Получить согласие по ID",
          "description": "## 📋 Получение согласия на управление договорами",
          "operationId": "get_product_agreement_consent_product_agreement_consents__consent_id__get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "consent_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Consent Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "title": "Response Get Product Agreement Consent Product Agreement Consents  Consent Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "6 Согласия на управление договорами"
          ],
          "summary": "3. Отозвать согласие",
          "description": "## 🗑️ Отзыв согласия на управление договорами",
          "operationId": "revoke_product_agreement_consent_product_agreement_consents__consent_id__delete",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "consent_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Consent Id"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful Response"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "CloseAgreementRequest": {
          "properties": {
            "repayment_account_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repayment Account Id"
            },
            "repayment_amount": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repayment Amount"
            }
          },
          "type": "object",
          "title": "CloseAgreementRequest",
          "description": "Запрос на закрытие договора с погашением"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ProductAgreementConsentRequestData": {
          "properties": {
            "requesting_bank": {
              "type": "string",
              "title": "Requesting Bank",
              "description": "Код банка-инициатора"
            },
            "client_id": {
              "type": "string",
              "title": "Client Id",
              "description": "ID клиента"
            },
            "read_product_agreements": {
              "type": "boolean",
              "title": "Read Product Agreements",
              "description": "Читать список договоров",
              "default": false
            },
            "open_product_agreements": {
              "type": "boolean",
              "title": "Open Product Agreements",
              "description": "Открывать новые договоры",
              "default": false
            },
            "close_product_agreements": {
              "type": "boolean",
              "title": "Close Product Agreements",
              "description": "Закрывать договоры",
              "default": false
            },
            "allowed_product_types": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Allowed Product Types",
              "description": "Разрешенные типы продуктов: deposit, card, credit"
            },
            "max_amount": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Amount",
              "description": "Макс сумма открытия продукта"
            },
            "valid_until": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valid Until",
              "description": "Действует до"
            },
            "reason": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason",
              "description": "Причина запроса"
            }
          },
          "type": "object",
          "required": [
            "requesting_bank",
            "client_id"
          ],
          "title": "ProductAgreementConsentRequestData",
          "description": "Запрос на создание согласия для управления договорами"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "HTTPBearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
...
  }